rename class dir
