+++ src/core/class/Error.ts
